+++ workspace/apps/server/scripts/dev.ts
@@ -20,17 +20,45 @@ import {
     type ChildProcessWithoutNullStreams,
 } from "node:child_process";
 import chokidar from "chokidar";
+import chalk from "chalk";
+import deshake from "@repo/utils/misc/deshake";
 
+/**
+ * 开发服务子进程
+ */
 class DevSubProcess {
+    /**
+     * 服务子进程
+     */
     private _subprocess?: ChildProcessWithoutNullStreams;
 
-    constructor() {}
+    /**
+     * 是否子进程是否已退出
+     */
+    private _exited: Promise<number | null> = Promise.resolve(null);
+
+    /**
+     * 延时重启服务子进程 (消抖)
+     */
+    public readonly delayedRestart!: typeof this.restart;
+
+    /**
+     * @param _restart_delay_time 重启服务的延时时间 (单位: 毫秒 ms)
+     */
+    constructor(
+        private readonly _restart_delay_time = 1_000, //
+    ) {
+        this.delayedRestart = deshake(this.restart.bind(this), this._restart_delay_time);
+    }
 
     public get subprocess(): ChildProcessWithoutNullStreams | undefined {
         return this._subprocess;
     }
 
-    async start() {
+    /**
+     * 启动服务子进程
+     */
+    public async start() {
         await this.stop();
         if (!this._subprocess) {
             try {
@@ -44,11 +72,33 @@ class DevSubProcess {
                         env: process.env,
                     },
                 );
-                this._subprocess.stdout?.on("data", (chunk) => console.log(chunk.toString()));
-                this._subprocess.stderr?.on("data", (chunk) => console.warn(chunk.toString()));
-                this._subprocess.on("error", (chunk) => console.error(chunk.toString()));
+                this._exited = new Promise<number | null>((resolve) => {
+                    if (this._subprocess) {
+                        this._subprocess.on("exit", (code) => {
+                            console.info(
+                                `\nSubprocess exit with code: ${(() => {
+                                    switch (code) {
+                                        case null:
+                                            return chalk.bgCyan(code);
+                                        case 0:
+                                            return chalk.bgGreen(code);
+                                        default:
+                                            return chalk.bgRed(code);
+                                    }
+                                })()}`,
+                            );
+                            resolve(code);
+                        });
+                    } else {
+                        resolve(null);
+                    }
+                });
 
-                console.info(`Subprocess run at PID: ${this._subprocess.pid}`);
+                this._subprocess.stdout?.on("data", process.stdout.write.bind(process.stdout));
+                this._subprocess.stderr?.on("data", process.stderr.write.bind(process.stderr));
+                this._subprocess.on("error", process.stderr.write.bind(process.stderr));
+
+                console.info(`Subprocess run at PID: ${chalk.blue(this._subprocess.pid)}\n`);
             } catch (error) {
                 console.error(error);
                 this._subprocess = undefined;
@@ -56,21 +106,14 @@ class DevSubProcess {
         }
     }
 
-    async stop() {
+    /**
+     * 终止服务子进程
+     */
+    public async stop() {
         if (this._subprocess) {
             try {
-                const killed = new Promise((resolve) => {
-                    if (this._subprocess) {
-                        this._subprocess.on("exit", (code) => {
-                            console.info(`Subprocess exit with code: ${code}`);
-                            resolve(undefined);
-                        });
-                    } else {
-                        resolve(undefined);
-                    }
-                });
                 this._subprocess.kill();
-                await killed;
+                await this._exited;
             } catch (error) {
                 console.error(error);
             } finally {
@@ -79,12 +122,18 @@ class DevSubProcess {
         }
     }
 
-    async restart() {
+    /**
+     * 重启服务子进程
+     */
+    public async restart() {
         await this.stop();
         await this.start();
     }
 }
 
+/**
+ * 开发服务
+ */
 async function dev() {
     const devSubProcess = new DevSubProcess();
 
@@ -109,12 +158,12 @@ async function dev() {
                 stats,
             ) => {
                 // console.log(eventName);
-                await devSubProcess.restart();
+                devSubProcess.delayedRestart();
             },
         );
     });
 }
 
-if (process.argv.at(-1) === import.meta.filename) {
+if (process.argv.includes(import.meta.filename)) {
     dev();
 }