+++ workspace/apps/server/tests/trpc/test.test.ts
@@ -16,7 +16,8 @@
  */
 
 import {
-    describe, //
+    //
+    describe,
     test,
     expect,
 } from "@jest/globals";