+++ workspace/packages/utils/crypto/index.ts
@@ -15,4 +15,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-console.log(import.meta.filename);
+export * from "./answer";
+export * from "./converter";
+export * from "./key";